requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
arxiv>=2.1.0
tavily-python>=0.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
mlflow>=2.10.0
streamlit>=1.32.0
plotly>=5.18.0
requests>=2.31.0
loguru>=0.7.0
pyyaml>=6.0
pytest>=7.4.0
ruff>=0.3.0
